# openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Fonoteca.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/catalogue/artists/{id}/portrait": {
      "get": {
        "tags": [
          "Catalogue"
        ],
        "summary": "The artist's picture, from the shelf where one exists.",
        "description": "An uploaded picture first, then `artist.*` beside the artist's own records, then a redirect to whatever a provider found. The shelf outranks the provider because it is what every other player reading this library shows, and the tag write is what puts a picture there. Served under an ETag with `no-cache`, so a changed picture shows on the next view.",
        "operationId": "GetArtistPortrait",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "width",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "302": {
            "description": "Found"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Catalogue"
        ],
        "summary": "Use an uploaded picture for this artist.",
        "description": "The request body is the image itself, its Content-Type one of JPEG, PNG, GIF, WebP, BMP or AVIF — never SVG, which is a document that runs script. At most 10 MB. Stored beside the provider's answer rather than over it, so deleting it puts the provider's picture back. It reaches the library the next time tags are written for this artist.",
        "operationId": "UploadArtistPortrait",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Catalogue"
        ],
        "summary": "Stop using the uploaded picture for this artist.",
        "description": "The provider's picture applies again. The copy already written onto the artist's shelf is left where it is until tags are written again — this endpoint touches the catalogue, never the library.",
        "operationId": "DeleteArtistPortrait",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/catalogue/artists/{id}/banner": {
      "get": {
        "tags": [
          "Catalogue"
        ],
        "summary": "The wide picture for the head of the artist's page.",
        "description": "The same three sources the portrait has, reading `backdrop.*` off the shelf. Navidrome does not show artist banners at all; Jellyfin, Kodi and Plex do.",
        "operationId": "GetArtistBanner",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "width",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "302": {
            "description": "Found"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Catalogue"
        ],
        "summary": "Use an uploaded picture as this artist's banner.",
        "description": "The request body is the image itself, its Content-Type one of JPEG, PNG, GIF, WebP, BMP or AVIF — never SVG, which is a document that runs script. At most 10 MB. Stored beside the provider's answer rather than over it, so deleting it puts the provider's picture back. It reaches the library the next time tags are written for this artist.",
        "operationId": "UploadArtistBanner",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Catalogue"
        ],
        "summary": "Stop using the uploaded banner for this artist.",
        "description": "The provider's picture applies again. The copy already written onto the artist's shelf is left where it is until tags are written again — this endpoint touches the catalogue, never the library.",
        "operationId": "DeleteArtistBanner",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    }
  },
...
}